refactor(appmenu): move the duplicate logic into hook
