+++ src/components/appMenu/CircularMenu/CircularMenuItem.module.scss
@@ -1,6 +1,6 @@
 @import '../../../style/mixins.scss';
 
-.menu_item {
+.menuItem {
   box-shadow: 0px 0px 13px #ffffff54;
   opacity: 1;
   background-color: #101010;
@@ -13,14 +13,14 @@
   pointer-events: auto;
 }
 
-.menu_item.active {
+.menuItem.active {
   border: 1px solid #ffffff6b;
   box-shadow: 0px 0px 8px #ffffff98 inset, 0px 0px 13px #ffffff98;
   width: 100%;
   height: 100%;
 }
 
-.menu_item:hover {
+.menuItem:hover {
   border: 1px solid #4ed6ae;
   box-shadow: 0px 0px 8px #4ed6ae inset, 0px 0px 13px #4ed6ae;
 }
